+++ src/coins.ts
@@ -146,7 +146,7 @@ const basicCoins: BasicCoin[] = [
         address: '0x0000000000000000000000000000000000000000',
         decimals: 18,
       },
-      [ChainId.CZKT]: {
+      [ChainId.LNAT]: {
         address: '0x0000000000000000000000000000000000000000',
         decimals: 18,
       },
@@ -724,10 +724,10 @@ const basicCoins: BasicCoin[] = [
         address: '0x6d4dd09982853f08d9966ac3ca4eb5885f16f2b2',
         decimals: 6,
       },
-      [ChainId.CZKT]: {
+      [ChainId.LNAT]: {
         address: '0x964FF70695da981027c81020B1c58d833D49A640',
         decimals: 6,
-        name: 'ConsenSys USD Coin',
+        name: 'Linea USD Coin',
       },
       // 42, 0xb7a4f3e9097c08da09517b5ab877f7a917224ede, 6
     },
@@ -813,6 +813,10 @@ const basicCoins: BasicCoin[] = [
       },
       [ChainId.OPTT]: {
         address: '0xe71678794fff8846bff855f716b0ce9d9a78e844',
+        decimals: 18,
+      },
+      [ChainId.LNAT]: {
+        address: '0xb706319d37b945727e71ae0d4353699d19112576',
         decimals: 18,
       },
     },
@@ -1019,7 +1023,7 @@ const basicCoins: BasicCoin[] = [
         address: '0x4200000000000000000000000000000000000006',
         decimals: 18,
       },
-      [ChainId.CZKT]: {
+      [ChainId.LNAT]: {
         address: '0x2C1b868d6596a18e32E61B901E4060C872647b6C',
         decimals: 18,
       },
@@ -1320,12 +1324,6 @@ const basicCoins: BasicCoin[] = [
         decimals: 9,
         symbol: 'WSOL',
         name: 'Token Wrapped SOL (Wormhole)',
-      },
-
-      // Testnet
-      [ChainId.SOL]: {
-        address: 'So11111111111111111111111111111111111111112',
-        decimals: 9,
       },
     },
   },
@@ -1697,12 +1695,12 @@ export const wrappedTokens: { [ChainId: string]: Token } = {
     logoURI:
       'https://zapper.fi/images/networks/binance-smart-chain/0x0000000000000000000000000000000000000000.png',
   },
-  [ChainId.CZKT]: {
+  [ChainId.LNAT]: {
     // https://explorer.prealpha.zkevm.consensys.net/address/0x2C1b868d6596a18e32E61B901E4060C872647b6C
     address: '0x2C1b868d6596a18e32E61B901E4060C872647b6C',
     symbol: 'WETH',
     decimals: 18,
-    chainId: ChainId.CZKT,
+    chainId: ChainId.LNAT,
     coinKey: CoinKey.WETH,
     name: 'WETH',
     logoURI: