+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [3.2.3](https://github.com/lifinance/types/compare/v3.2.2...v3.2.3) (2023-04-13)
+
+
+### Bug Fixes
+
+* add Linea Test token ([#157](https://github.com/lifinance/types/issues/157)) ([c448503](https://github.com/lifinance/types/commit/c448503bd84db9aefc8b1d4275950764fbff03a9))
+
+### [3.2.2](https://github.com/lifinance/types/compare/v3.2.1...v3.2.2) (2023-04-12)
+
+
+### Bug Fixes
+
+* rename CZKT to LNA (Linea) ([#156](https://github.com/lifinance/types/issues/156)) ([b414277](https://github.com/lifinance/types/commit/b414277a37ebeefa37871f2e766915d9d75dd65b))
+
+### [3.2.1](https://github.com/lifinance/types/compare/v3.2.0...v3.2.1) (2023-04-11)
+
+
+### Bug Fixes
+
+* step types cleanup ([#155](https://github.com/lifinance/types/issues/155)) ([8d3caed](https://github.com/lifinance/types/commit/8d3caedc417bdcc3e4fedba575f1438092be8f23))
+
 ## [3.2.0](https://github.com/lifinance/types/compare/v3.1.1...v3.2.0) (2023-04-05)